pyproject.toml
[project]
name = "pyya"
version = "0.1.7"
authors = [
  { name = "shadowy-pycoder", email = "shadowy-pycoder@example.com" },
]
description = "Convert YAML configuration files to Python objects"
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "License :: OSI Approved :: MIT License",
  "Operating System :: MacOS :: MacOS X",
  "Operating System :: Microsoft :: Windows",
  "Operating System :: POSIX :: Linux",
]
dependencies = [
  "camel-converter>=3.1.2",
  "munch>=4.0.0",
  "pydantic>=2.5.2",
  "pyyaml>=6.0.2",
  "types-pyyaml>=6.0.12.20240917",
]

[project.urls]
Homepage = "https://github.com/shadowy-pycoder/pyya"
Repository = "https://github.com/shadowy-pycoder/pyya"
Issues = "https://github.com/shadowy-pycoder/pyya/issues"

[tool.setuptools]
license-files = []

[tool.mypy]
python_version = "3.8"
cache_dir = ".mypy_cache/strict"
allow_redefinition = true
strict_optional = false
show_error_codes = true
show_column_numbers = true
warn_no_return = true
disallow_any_unimported = false
warn_unused_configs = true
disallow_any_generics = true
disallow_subclassing_any = false
disallow_untyped_calls = false
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
implicit_reexport = true
strict_equality = true
ignore_missing_imports = true

[tool.pytest.ini_options]
pythonpath = "."
filterwarnings = [
  "ignore::DeprecationWarning",
  "ignore::PendingDeprecationWarning",
  "ignore::ImportWarning",
  "ignore::ResourceWarning",
]
asyncio_mode = "auto"
env_files = ".test.env"
env_override_existing_values = true

[tool.ruff]
exclude = [
  ".bzr",
  ".direnv",
  ".eggs",
  ".git",
  ".git-rewrite",
  ".hg",
  ".ipynb_checkpoints",
  ".mypy_cache",
  ".nox",
  ".pants.d",
  ".pyenv",
  ".pytest_cache",
  ".pytype",
  ".ruff_cache",
  ".svn",
  ".tox",
  ".venv",
  ".vscode",
  "__pypackages__",
  "_build",
  "buck-out",
  "build",
  "dist",
  "node_modules",
  "site-packages",
  "venv",
]
line-length = 120
indent-width = 4
target-version = "py38"

[tool.ruff.format]
quote-style = "single"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.lint]
select = ["E1", "E2", "E4", "E7", "E9", "F", "D3"]
ignore = ["ANN201", "Q000", "D102", "D103"]
fixable = ["ALL"]
unfixable = []
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.lint.isort]
lines-after-imports = 2
